refactor: 숫자인지 확인 하는 방법 변경 및 뷰들 폴더로 묶기
