+++ src/main/java/calculator/view/InputView.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view;
 
 public class InputView {
     public InputView() {
